+++ async_writer.go
@@ -29,27 +29,27 @@ type asyncWriter struct {
 
 // AsyncWriterOption represents available options for async writer.
 type AsyncWriterOption struct {
-	Ref       Writer
+	RefWriter Writer
 	QueueSize int
 }
 
 // NewAsyncWriter creates a new instance of asynchronous writer.
 func NewAsyncWriter(options ...func(*AsyncWriterOption)) Writer {
-	opt := &AsyncWriterOption{
+	opts := &AsyncWriterOption{
 		QueueSize: defaultWriterQueueSize,
 	}
 
 	for _, f := range options {
-		f(opt)
+		f(opts)
 	}
 
-	if opt.Ref == nil {
+	if opts.RefWriter == nil {
 		ReportfExit("async writer need a referenced writer")
 	}
 
 	return &asyncWriter{
-		ref:   opt.Ref,
-		queue: NewBlockingQueue(opt.QueueSize),
+		ref:   opts.RefWriter,
+		queue: NewBlockingQueue(opts.QueueSize),
 	}
 }